+++ cc-gl/src/main/java/opengl4experiments/bareboneApp/BareboneGLExperimentWithCamera.java
@@ -71,7 +71,7 @@ public class BareboneGLExperimentWithCamera extends BareboneGLAbstractExperiment
 			int w, int h) {
 		width = w;
 		height = h;
-		aspect = (float)height/width;
+		aspect = height /width;
 		myCam.updateProjectionMatrix(aspect);
 	}
