+++ cc-image/src/main/java/cc/creativecomputing/image/filter/CCImageMath.java
@@ -408,7 +408,7 @@ public class CCImageMath {
 				break;
 		if (span > numKnots-3)
 			span = numKnots-3;
-		float t = (float)(x-xknots[span]) / (xknots[span+1]-xknots[span]);
+		float t = (x-xknots[span]) / (xknots[span+1]-xknots[span]);
 		span--;
 		if (span < 0) {
 			span = 0;
@@ -554,7 +554,7 @@ public class CCImageMath {
 		for (j = 0; j < length; j++) {
 			while (out[i+1] < j)
 				i++;
-			in[j] = i + (float) (j - out[i]) / (out[i + 1] - out[i]);
+			in[j] = i + (j - out[i]) / (out[i + 1] - out[i]);
 //			in[j] = CCImageMath.clamp( in[j], 0, length-1 );
 		}
 		in[length] = length;
